algorithms/BoundingBoxOnImage/requirements.txt
algorithmia>=1.0.0,<2.0
six
pillow>=8.0,<8.1

